Restructured 'asset' package.
