+++ game-engine/src/main/java/gameengine/game/Material.java
@@ -2,7 +2,7 @@ package gameengine.game;
 
 import org.joml.Vector4f;
 
-import gameengine.engine.asset.Texture;
+import gameengine.engine.asset.texture.Texture;
 
 public class Material {
 	public static final int DEFAULT_TEXTURE_SLOT_COUNT = 3;
