+++ game-engine/src/main/java/gameengine/game/Model.java
@@ -5,8 +5,8 @@ import java.util.List;
 import org.lwjgl.opengl.GL46;
 
 import gameengine.engine.IRenderable;
+import gameengine.engine.asset.mesh.Mesh;
 import gameengine.engine.renderer.IRenderPass;
-import gameengine.engine.renderer.Mesh;
 
 public class Model implements IRenderable {
     private List<Mesh> meshList;
