+++ game-engine/src/main/java/gameengine/game/TestAnother.java
@@ -11,9 +11,9 @@ import org.joml.Vector4f;
 
 import gameengine.engine.Engine;
 import gameengine.engine.IGameObject;
-import gameengine.engine.asset.Texture;
+import gameengine.engine.asset.mesh.Mesh;
+import gameengine.engine.asset.texture.Texture;
 import gameengine.engine.renderer.IRenderPass;
-import gameengine.engine.renderer.Mesh;
 import gameengine.util.FileUtils;
 
 public class TestAnother implements IGameObject {
